Fix 64-bit build support.
- This fixes some typecasting issues that broke 64-bit support.
- This fixes some misdefined macros in 'platform.h' that didn't work right in 64-bit mode.
- This fixes the inability to use '__asm' in 64-bit mode.

All unit tests pass in the 64-bit debug build again.

+++ smilelib/src/eval/bytecode.c
@@ -44,7 +44,7 @@ ByteCodeSegment ByteCodeSegment_CreateWithSize(Int size)
 
 	segment->byteCodes = byteCodes;
 	segment->numByteCodes = 0;
-	segment->maxByteCodes = size;
+	segment->maxByteCodes = (Int32)size;
 
 	return segment;
 }
@@ -72,7 +72,7 @@ void ByteCodeSegment_Grow(ByteCodeSegment segment, Int count)
 	MemCpy(newByteCodes, segment->byteCodes, sizeof(struct ByteCodeStruct) * segment->numByteCodes);
 
 	segment->byteCodes = newByteCodes;
-	segment->maxByteCodes = newMax;
+	segment->maxByteCodes = (Int32)newMax;
 }
 
 STATIC_STRING(_colonString, ":");
